requirements.txt
python-telegram-bot==20.7
flask==2.3.3
flask-login==0.6.2
flask-jwt-extended==4.5.2
python-dotenv==1.0.0
gspread==5.12.0
oauth2client==4.1.3
openpyxl==3.1.2
aiofiles==23.2.1
loguru==0.7.2
requests==2.31.0
aiohttp==3.8.5